welness/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-auditlog==0.4.5
django-cors-headers==2.2.0
django-jsonfield==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
future==0.16.0
gevent==1.3.0
greenlet==0.4.13
grequests==0.3.0
gunicorn==19.7.1
idna==2.6
natsort==5.1.1
olefile==0.44
pexpect==4.3.0
Pillow==4.3.0
psycopg2==2.7.3.2
ptyprocess==0.5.2
PyJWT==1.6.4
python-dateutil==2.6.0
python-snappy==0.5.2
pytz==2017.3
pyzmq==16.0.3
requests==2.18.4
six==1.11.0
sysv-ipc==1.0.0
urllib3==1.22
zmq==0.0.0
